refactor(-): :hammer: refatora rotas configuracoes, GA e DE
